# apps/api/internal/handlers/openapi.yaml
openapi: 3.1.0
info:
  title: BrokeStack API
  version: 0.1.0
  description: |
    BrokeStack is a multi-tenant platform for automation, utility APIs,
    scraping workflows, AI-assisted tools, and operator-managed workspaces.
    
    ## Platform Areas
    - Customer workspaces and API access
    - Tool execution and job orchestration
    - Admin and support operations
    - Authentication, tenancy, and permissions
    - Future Python worker integration for long-running tasks
    
    ## Authentication
    Most endpoints require authentication via Bearer token in the Authorization header.
  contact:
    name: BrokeStack API Support
    email: api@brokestack.app
  license:
    name: MIT
    identifier: MIT

servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.memofi.tech
    description: Production server

tags:
  - name: Health
    description: Health check and status endpoints
  - name: Images
    description: Browser-backed image generation and image history
  - name: Workspace API
    description: Workspace API keys, quotas, and docs
  - name: Auth
    description: Authentication and authorization
  - name: Workspaces
    description: Multi-tenant workspace management
  - name: Jobs
    description: Tool execution and background run lifecycle
  - name: Tooling
    description: Tool catalog, usage, and future worker integration
  - name: Platform
    description: Admin and workspace platform capabilities

paths:
  /api/v1/health:
    get:
      operationId: getHealth
      summary: Health Check
      description: Returns the health status of the API including uptime and system information.
      tags:
        - Health
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: healthy
                timestamp: "2026-03-06T12:00:00Z"
                uptime: "2h30m15s"
                version: "0.1.0"
                go:
                  version: "go1.24"
                  os: "darwin"
                  arch: "arm64"
                  num_cpu: 10
                  goroutines: 8
                checks:
                  api: "ok"
  /api/v1/images:
    post:
      operationId: createImageJob
      summary: Create image job
      description: Queue a single image generation job using a workspace API key.
      tags:
        - Images
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateImageJobRequest'
      responses:
        '201':
          description: Image job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageJob'
        '4XX':
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/image-batches:
    post:
      operationId: createImageBatch
      summary: Create image batch
      description: Queue a grouped batch request that fans out into individually scheduled image jobs.
      tags:
        - Images
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateImageBatchRequest'
      responses:
        '201':
          description: Image batch created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageBatchResult'
        '4XX':
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/images/{id}:
    get:
      operationId: getImageJob
      summary: Get image job
      description: Fetch the current state of one image job created through the API or dashboard.
      tags:
        - Images
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Image job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageJob'
        '4XX':
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/images/{id}/outputs:
    get:
      operationId: listImageOutputs
      summary: List image outputs
      description: List the downloaded outputs for a completed image job.
      tags:
        - Images
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Image outputs
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ImageOutput'
        '4XX':
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/workspaces/{workspaceId}/api-keys:
    get:
      operationId: listWorkspaceApiKeys
      summary: List workspace API keys
      tags:
        - Workspace API
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Workspace API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkspaceApiKey'
    post:
      operationId: createWorkspaceApiKey
      summary: Create workspace API key
      tags:
        - Workspace API
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceApiKeyRequest'
      responses:
        '201':
          description: Workspace API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedWorkspaceApiKey'
        '4XX':
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token

  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - timestamp
        - version
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
          description: Overall health status
        timestamp:
          type: string
          format: date-time
          description: Current server timestamp
        uptime:
          type: string
          description: Server uptime duration
        version:
          type: string
          description: API version
        go:
          $ref: '#/components/schemas/GoInfo'
        checks:
          type: object
          additionalProperties:
            type: string
          description: Individual service health checks

    GoInfo:
      type: object
      properties:
        version:
          type: string
          description: Go runtime version
        os:
          type: string
          description: Operating system
        arch:
          type: string
          description: CPU architecture
        num_cpu:
          type: integer
          description: Number of CPUs
        goroutines:
          type: integer
          description: Active goroutines count

    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
          description: Error message
        code:
          type: string
          description: Error code for programmatic handling
        details:
          type: object
          additionalProperties: true
          description: Additional error details

    CreateImageJobRequest:
      type: object
      required:
        - promptText
      properties:
        title:
          type: string
        promptText:
          type: string
        aspectRatio:
          type: string

    CreateImageBatchRequest:
      type: object
      required:
        - promptTemplate
        - values
      properties:
        title:
          type: string
        promptTemplate:
          type: string
        placeholderName:
          type: string
        aspectRatio:
          type: string
        values:
          type: array
          items:
            type: string

    ImageJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        providerAccountId:
          type: string
          format: uuid
        provider:
          type: string
        providerLabel:
          type: string
        batchId:
          type: string
          format: uuid
        apiKeyId:
          type: string
          format: uuid
        source:
          type: string
          enum: [web, api]
        requestType:
          type: string
          enum: [single, batch]
        title:
          type: string
        promptText:
          type: string
        aspectRatio:
          type: string
        status:
          type: string
        queuedAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        lastError:
          type: string
        outputCount:
          type: integer

    ImageOutput:
      type: object
      properties:
        id:
          type: string
          format: uuid
        imageJobId:
          type: string
          format: uuid
        mimeType:
          type: string
        byteSize:
          type: integer
        width:
          type: integer
        height:
          type: integer
        contentUrl:
          type: string

    ImageBatch:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        providerAccountId:
          type: string
          format: uuid
        title:
          type: string
        promptTemplate:
          type: string
        placeholderName:
          type: string
        source:
          type: string
        status:
          type: string

    ImageBatchResult:
      type: object
      properties:
        batch:
          $ref: '#/components/schemas/ImageBatch'
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/ImageJob'

    WorkspaceApiKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        name:
          type: string
        keyPrefix:
          type: string
        status:
          type: string
        scopes:
          type: array
          items:
            type: string
        requestsPerMinute:
          type: integer
        dailyImageQuota:
          type: integer
        lastUsedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateWorkspaceApiKeyRequest:
      type: object
      required:
        - name
        - scopes
        - requestsPerMinute
        - dailyImageQuota
      properties:
        name:
          type: string
        scopes:
          type: array
          items:
            type: string
        requestsPerMinute:
          type: integer
        dailyImageQuota:
          type: integer

    CreatedWorkspaceApiKey:
      type: object
      properties:
        apiKey:
          $ref: '#/components/schemas/WorkspaceApiKey'
        secret:
          type: string